all: clippy lint and fix tests

+++ libbpf-rs/tests/test.rs
@@ -127,7 +127,7 @@ fn test_object_percpu_lookup() {
     let map = obj.map_mut("percpu_map").expect("failed to find map");
 
     let res = map
-        .lookup_percpu(&(0 as u32).to_ne_bytes(), MapFlags::ANY)
+        .lookup_percpu(&(0_u32).to_ne_bytes(), MapFlags::ANY)
         .expect("failed to lookup")
         .expect("failed to find value for key");
 
@@ -145,9 +145,7 @@ fn test_object_percpu_invalid_lookup_fn() {
     let mut obj = get_test_object("percpu_map.bpf.o");
     let map = obj.map_mut("percpu_map").expect("failed to find map");
 
-    assert!(map
-        .lookup(&(0 as u32).to_ne_bytes(), MapFlags::ANY)
-        .is_err());
+    assert!(map.lookup(&(0_u32).to_ne_bytes(), MapFlags::ANY).is_err());
 }
 
 #[test]
@@ -156,7 +154,7 @@ fn test_object_percpu_update() {
 
     let mut obj = get_test_object("percpu_map.bpf.o");
     let map = obj.map_mut("percpu_map").expect("failed to find map");
-    let key = (0 as u32).to_ne_bytes();
+    let key = (0_u32).to_ne_bytes();
 
     let mut vals: Vec<Vec<u8>> = Vec::new();
     for i in 0..num_possible_cpus().unwrap() {
@@ -180,9 +178,9 @@ fn test_object_percpu_invalid_update_fn() {
 
     let mut obj = get_test_object("percpu_map.bpf.o");
     let map = obj.map_mut("percpu_map").expect("failed to find map");
-    let key = (0 as u32).to_ne_bytes();
+    let key = (0_u32).to_ne_bytes();
 
-    let val = (1 as u32).to_ne_bytes().to_vec();
+    let val = (1_u32).to_ne_bytes().to_vec();
 
     assert!(map.update(&key, &val, MapFlags::ANY).is_err());
 }
@@ -193,7 +191,7 @@ fn test_object_percpu_lookup_update() {
 
     let mut obj = get_test_object("percpu_map.bpf.o");
     let map = obj.map_mut("percpu_map").expect("failed to find map");
-    let key = (0 as u32).to_ne_bytes();
+    let key = (0_u32).to_ne_bytes();
 
     let mut res = map
         .lookup_percpu(&key, MapFlags::ANY)
